Moved dropdown close to the tag being typed
The tag list dropdown used to always appear at the bottom of the
text area; this is undesirable if the text area gets big and you
are editing a tag at the top of the text.

These changes move the dropdown close to the tag being edited
by the user.

This code is a bit kludgy and may require further tweaking, but
it is a good start.